fix: comment settings
