+++ src/components/data/Fieldset/Fieldset.stories.tsx
@@ -25,12 +25,6 @@ const meta: Meta<typeof Fieldset> = {
     children: {
       description: "children of the component",
     },
-    isMobile: {
-      control: {
-        type: "boolean",
-      },
-      description: "isMobile of the component",
-    },
   },
 };
