+++ src/components/data/Fieldset/index.tsx
@@ -2,14 +2,33 @@ import { useState } from "react";
 import { MdAdd, MdOutlineMoreVert } from "react-icons/md";
 import { Stack, Text, useMediaQuery, Button, Icon } from "@inubekit/inubekit";
 
+import { spacing } from "@design/tokens/spacing";
+
 import {
   StyledContainerFieldset,
   StyledMenuContainer,
   StyledPrint,
 } from "./styles";
-import { FieldsetProps } from "./types";
+import { IOptionsButton } from "./types";
 
-export const Fieldset = (props: FieldsetProps) => {
+export interface FieldsetProps {
+  children: React.ReactNode;
+  title: string;
+  descriptionTitle?: string;
+  aspectRatio?: string;
+  heightFieldset?: string;
+  activeButton?: IOptionsButton;
+  hasTable?: boolean;
+  hasOverflow?: boolean;
+  isClickable?: boolean;
+  selectedState?: boolean;
+  hasError?: boolean;
+  borderColor?: "normal" | "error" | "success";
+  alignContent?: "left" | "center" | "right";
+  onSelectionChange?: () => void;
+}
+
+export function Fieldset(props: FieldsetProps) {
   const {
     children,
     title,
@@ -41,12 +60,12 @@ export const Fieldset = (props: FieldsetProps) => {
   return (
     <Stack
       direction="column"
-      gap="8px"
+      gap={spacing.s100}
       width="-webkit-fill-available"
       height={!isMobile ? heightFieldset : "auto"}
     >
       <Stack justifyContent={activeButton && "space-between"}>
-        <Stack gap={isMobile ? "12px" : "8px"}>
+        <Stack gap={isMobile ? spacing.s150 : spacing.s100}>
           <Text type="title" size={isMobile ? "medium" : "large"}>
             {title}
           </Text>
@@ -112,4 +131,4 @@ export const Fieldset = (props: FieldsetProps) => {
       </StyledContainerFieldset>
     </Stack>
   );
-};
+}
